.claude/edits/20260301_221121_requirements.txt
twikit>=2.3.3
gspread>=6.1.4
google-auth-oauthlib>=1.2.1
google-auth>=2.38.0
